DDF-852  added check for empty list of metacard IDs before attempting Solr cache deletion - added to prevent stack trace complaining about missing content stream when trying to delte non-existent metacard

+++ core/catalog-core-standardframework/src/main/java/ddf/catalog/cache/solr/impl/SolrCache.java
@@ -442,8 +442,12 @@ public class SolrCache {
                 for (Metacard deletedMetacard : deletedMetacards) {
                     metacardIdentfiers.add(deletedMetacard.getId());
                 }
-                LOGGER.debug("metacard identifiers to be deleted: " + StringUtils.join(metacardIdentfiers, ","));
-                server.deleteById(metacardIdentfiers);
+                if (!metacardIdentfiers.isEmpty()) {
+                    LOGGER.debug("metacard identifiers to be deleted: " + StringUtils.join(metacardIdentfiers, ","));
+                    server.deleteById(metacardIdentfiers);
+                } else {
+                    LOGGER.debug("No metacard identifiers to be deleted");
+                }
             }
         } catch (SolrServerException e) {
             LOGGER.error("SOLR server exception deleting request message", e);
